% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outbreak_control.R
\name{plot_outbreak_control}
\alias{plot_outbreak_control}
\title{Plot the percentage of outbreaks controlled again the percentage of
contacts traced for various outbreak parameters}
\usage{
plot_outbreak_control(data, var)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{var}{The variable that is being plotted, i.e. which variable is being
split by colour.}
}
\value{
A `ggplot` object.
}
\description{
Plot the percentage of outbreaks controlled again the percentage of
contacts traced for various outbreak parameters
}
